==> 2017_06_track_matlab/00_create_metadata.Rmd <==
---
title: "2017_04_28_metadata"
author: "TB"
date: "4/28/2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r cars}
library("tidyverse")
library("magrittr")
```
# Collect MATLAB tracking data 
The segmentation results and features of the matlab tracking are collected as a single data frame and combined with the metadata stored in a separate file. 

One big data frame is created and stored in the analysis folder. 

## Input

The matlab tracking data for each of the 94 experiments has its own folder that is stored in the file 'complete_metadata.csv'. In each folder, the single cell data can be found in a file called featureList.csv. 

```{r}
data_folder <- "../../../../2017_04_05_CP17/images/"
csv_identifier <- "featureList.csv"
metadata_file <- "../../../../../2017_03_08_neutrophile_migration_becker/2017_04_05_CP17/complete_metadata.csv"
```

Output 
The folder analysis_folder stores all results. Here 
```{r}
analysis_folder <- '../../../analysis/2017_04_28_matlab/'
csv_single_cell_data <- 'TrackeddetectedNeutros.csv'

metadata <- read_csv(metadata_file) %>%
  rename(Metadata_condition = Metadata_dose_1) %>%
  print()

experiment_list <- metadata %>%
  extract2("File_Path")

strata <- c('Metadata_condition','Metadata_dose','Metadata_id','Metadata_date','Metadata_matrix')
```
Metadata_date is missing, we add it using previously tracked data 
```{r}
df_temp <- read_csv(file.path(analysis_folder, 'normalized.csv')) %>% 
  select(Metadata_id, Metadata_matrix, Metadata_dose, Metadata_condition, Metadata_date) %>% 
  group_by_(.dots = strata) %>% 
  slice(1) %>%
  print

metadata %<>% 
  mutate(Metadata_condition = replace(Metadata_condition, Metadata_condition == 'cont', 'control')) %>% 
  mutate(Metadata_condition = replace(Metadata_condition, Metadata_condition == 'vehi', 'vehicle')) %>%
  left_join(.,df_temp) %>% print


```
We read all files s

```{r}
read_file <- function(path){
  read_csv(file.path(path,csv_identifier))
}

df_list <- map(experiment_list, read_file)
```

```{r}
df_list2 <- list()
for (i in 1:nrow(metadata)){
  df_list2[[i]] = cbind2(metadata %>% 
                            slice(rep(i, each = nrow(df_list[[i]]))),
                         df_list[[i]])
}
```

```{r}
df <- do.call(rbind, df_list2) 

write_csv(df, file.path(analysis_folder, csv_single_cell_data))

```
