==> 2017_06_track_matlab/02_LTB4_assess.Rmd <==
---
title: "load_data"
author: "TB"
date: "5/15/2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library("tidyverse")
```

# Load data 
The output of CellProfiler track module is loaded as CSV, tracked using cytominer::track and 
stored in the data folder as "track.csv". 
```{r}
csv_in <- 'clean_tracks.csv'
csv_out <- 'assessed.csv'
analysis_folder <- '/Users/tbecker/Documents/2017_05_10_neutrophile_inhibitor/workspace/analysis/2018_ltb4'
#
strata <- c('Metadata_condition','Metadata_dose','Metadata_id',
            'Metadata_date','Metadata_matrix','Metadata_chemokine')

postprocessed <- read_csv(file.path(analysis_folder,csv_in)) %>% 
  group_by_(.dots = strata)
```

```{r}
assess <- function(tracks, min_path_length = 19, strata) {
  track_info <- list(migrationminer::valid_observation_time(tracks, min_path_length),
    migrationminer::validate_tracks(tracks, min_path_length))

  return(Reduce(function(...) merge(..., all = TRUE, by_ = strata), track_info))
}


quality <- migrationminer::assess(postprocessed, min_path_length = 20)
```

## Assess quality
```{r}
result_gyde <- left_join(
  
  quality <- migrationminer::assess(postprocessed,min_path_length = 20),
  
  postprocessed %>%
    filter(Track_Length > 20) %>%
    summarise(
      tracks_in_positive_sector = sum(Track_Positive_Sector),
      tracks_in_negative_sector = sum(Track_Negative_Sector),
      tracks_in_neutral_up_sector = sum(Track_Neutral_Sector_Up),
      tracks_in_neutral_down_sector = sum(Track_Neutral_Sector_Down)),
  by = strata) %>%
  select(-sum_track) 


result_gyde %>% print


```

Data export for Gyde
Export mean VOT for all conditions / doses / matrices

```{r}
sem <- function(vector){
  sd(vector)
}

q1 <- quality %>% 

  ungroup() %>%
  group_by(Metadata_matrix, Metadata_condition, Metadata_dose) %>%
  summarise(
    n = n(),
    VOT = mean(VOT)
  ) %>% print

```
