==> 2017_06_track_matlab/05_analyse_inhibitor.Rmd <==
---
title: "load_data"
author: "TB"
date: "07/07/2017"
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library("readr")
library("tibble")
library("dplyr")
library("tidyr")
library("stringr")
library("ggplot2")
library("purrr")
library("broom")
library("dplyr")
library("tsne")
library("ggfortify")
library("cluster")
library("magrittr")
library("knitr")
library("corrplot")
source("../windrose.r")
library("destiny")
library("scatterplot3d")
library("car")
```


# Load data 
Load tracking data created using MATLAB as normalized and not normalized data. 

The normalization was performed using cytominer::normalize and the following variables 
are normalized: 
* "Track_Speed",
* "Track_Speed_Y",
* "Track_Speed_X",
* "Track_xFMI",
* "Track_yFMI",
* "Track_Directionality",
* "Track_Distance_Traveled", 
* "Track_CI"


```{r}
csv_tracks <- 'clean_tracks.csv'
csv_normalized <- 'clean_normalized.csv'
analysis_folder <- '../../../analysis/2017_04_28_matlab/'
strata <- c('Metadata_condition','Metadata_dose','Metadata_id','Metadata_date','Metadata_matrix')

tracks <- read_csv(file.path(analysis_folder, csv_tracks)) %>%
  select(Metadata_id,Metadata_matrix, Metadata_condition, Metadata_dose, everything() ) %>%
  arrange(Metadata_id, Metadata_matrix, Metadata_condition, Metadata_dose) %>%
  group_by_(.dots = strata) 

normalized <- read_csv(file.path(analysis_folder, csv_normalized)) %>%
  select(Metadata_id,Metadata_matrix, Metadata_condition, Metadata_dose, everything() ) %>%
  arrange(Metadata_id, Metadata_matrix, Metadata_condition, Metadata_dose) %>%
  group_by_(.dots = strata) 


```
Small trajectories are removed and the angles are scaled to [-pi, pi] and rotated 
(using cytominer::track the angles were calculated with 0 degree set to three o 
clock, now we rotate zero 12 )

Small trajectories of neutrophils that are observred less than 20 frames are removed. 
```{r}
# angles are scaled to the intervale (-pi, pi)
population <- tracks %>%
  filter(Track_Length > 9) %>%
  mutate(Track_Angle = Track_Angle + pi/2) %>% # rotate all angles by 90 degree or pi/2
  mutate(Track_Angle = ifelse(Track_Angle > pi, Track_Angle - 2*pi, Track_Angle) ) %>%
  mutate(Track_Angle = ifelse(Track_Angle < 0, Track_Angle + 2*pi, Track_Angle) ) 


```

## Quality 
To analyse the tracking quality we use cytominer::assess(). To compare the results of 
all migration run in FN and HEM we use the VOT (valid observation time). 

```{r}
quality <- tracks %>% 
  neutrominer::assess(min_path_length = 9)

quality %>% 
  ungroup() %>%
  group_by(Metadata_matrix) %>% 
  summarise(mean_VOT = mean(VOT), sd_VOT = sd(VOT)) %>% 
  print()
```

Lets visually compare the results from both matrices
```{r}
ggplot(data = quality %>% filter(Metadata_matrix == 'FN'), 
       aes(x = VOT, y = Exp_Valid_Tracks, color = Metadata_id, shape = Metadata_condition)) +
  geom_point() +
  coord_cartesian(xlim = c(0, 1)) +
  ggtitle("Track Quality for migration in FN (high VOT is good)")
  
ggplot(data = quality %>% filter(Metadata_matrix == 'HEM'), 
       aes(x = VOT, y = Exp_Valid_Tracks, color = Metadata_id, shape = Metadata_condition)) +
  geom_point() +
  coord_cartesian(xlim = c(0, 1)) + 
  ggtitle("Track Quality for migration in HEM (high VOT is good)")
```

Results: the migration in HEM shows a higher VOT compared to FN. Do we see a significant difference? Answer: yes, we get p = 0.0353 
```{r}
x <- quality %>% filter(Metadata_matrix == "FN")
y <- quality %>% filter(Metadata_matrix == "HEM")
t.test(x$VOT, y$VOT)
```

Question: Why is the tracking quality better in HEM? From previous analysis we know that the speed is different.
Note: for the following analysis we do not normalized data!

```{r}
population %>% 
  ungroup() %>% 
  #filter(Metadata_condition != "water") %>%
  group_by(Metadata_condition,Metadata_matrix ) %>%
  summarise(mean_speed = mean(Track_Speed), sd_speed = sd(Track_Speed))
```

The speed in FN is roughly twice as high as in HEM. With one exception: the CP17 migraions run in HEM. They also 
show a very high standard deviation. Hmm... 
```{r}
population %>% 
  ungroup() %>% 
  group_by(Metadata_id, Metadata_dose) %>%
  filter(Metadata_matrix == "HEM", Metadata_condition == "vehicle") %>%
  summarise(mean_speed = mean(Track_Speed), sd_speed = sd(Track_Speed))

```

We remove our outlier and look again. This looks much nicer :)
```{r}
population %>% 
  ungroup() %>% 
  group_by(Metadata_condition,Metadata_matrix ) %>%
  summarise(mean_speed = mean(Track_Speed), sd_speed = sd(Track_Speed))
```

Next we check if the speed is correlated to the tracking quality / VOT. 
```{r}
speed <- population %>% 
  summarise(speed = mean(Track_Speed)) %>%
  left_join(., quality, by = strata)

ggplot(data = speed, aes(x = speed, y = VOT, color = Metadata_matrix, shape = Metadata_condition)) +
  geom_point()

lin_reg <- lm(speed ~ VOT, data = speed) 
summary(lin_reg)
```
We get a slope of -5.2657, R^2 = 0.3. The linear regression shows a weak 
correlation between VOT and speed. 


## Migration analysis

No we want to look at the distribution of the direction and speed of the migration. 
We select the migrations run in Fibronectin. Each condition (CP17, SP17, control and vehicle) is plotted separately.

```{r}

test_data <- population %>%
  filter(Metadata_matrix == "HEM") 

CP17 <- test_data %>% 
  filter(Metadata_condition == "CP17") %>%
  filter(Metadata_dose == 50) 

CP17 %>% summarise(n = n())  

SP17 <- test_data %>% 
  filter(Metadata_condition == "SP17") %>%
  filter(Metadata_dose == 50) 
  
controls <- test_data %>% 
  filter(Metadata_condition == "control") 

vehicle <- test_data %>% 
  filter(Metadata_condition == "vehicle") 

spdmin <- 0
spdmax <- 16
spdres <- 1
dirres <- 90

# factor to convert speed from pixel / frame to mum / min

h1 <- plot.windrose(spd = CP17$Track_Speed, dir = (180 * (CP17$Track_Angle) / pi), 
  spdmin = spdmin, spdmax = spdmax, spdres = spdres, dirres = dirres, title_name = "Migration distribution CP17, all doses")

h2 <- plot.windrose(spd = SP17$Track_Speed, dir = (180 * (SP17$Track_Angle) / pi), 
  spdmin = spdmin, spdmax = spdmax, spdres = spdres, dirres = dirres, title_name = "Migration distribution SP17, all doses")

h3 <- plot.windrose(spd = controls$Track_Speed, dir = (180 * (controls$Track_Angle) / pi),
  spdmin = spdmin, spdmax = spdmax, spdres = spdres, dirres = dirres, title_name = "Migration distribution control)")

h4 <- plot.windrose(spd = vehicle$Track_Speed, dir = (180 * (vehicle$Track_Angle) / pi), 
  spdmin = spdmin, spdmax = spdmax, spdres = spdres, dirres = dirres,  title_name = "Migration distribution vehicle")
```
Using the following code we can create a "Finer"" plot for 12 sections.
```{r}
dirres <- 30

h1 <- plot.windrose(spd = CP17$Track_Speed, dir = (180 * (CP17$Track_Angle) / pi), 
  spdmin = spdmin, spdmax = spdmax, spdres = spdres, dirres = dirres, title_name = "Migration distribution CP17, all doses")

h2 <- plot.windrose(spd = SP17$Track_Speed, dir = (180 * (SP17$Track_Angle) / pi), 
  spdmin = spdmin, spdmax = spdmax, spdres = spdres, dirres = dirres, title_name = "Migration distribution SP17, all doses")

h3 <- plot.windrose(spd =  controls$Track_Speed, dir = (180 * (controls$Track_Angle) / pi),
  spdmin = spdmin, spdmax = spdmax, spdres = spdres, dirres = dirres, title_name = "Migration distribution control")

h4 <- plot.windrose(spd = vehicle$Track_Speed, dir = (180 * (vehicle$Track_Angle) / pi), 
  spdmin = spdmin, spdmax = spdmax, spdres = spdres, dirres = dirres,  title_name = "Migration distribution vehicle")
```
# Normalized data
For the further analysis, we use normalized data. The normalization was performed using cytominer::normalize, see **example_normalize_matlab.Rmd**. Again, we remove the bad experiment. 
```{r}
summary_speed <- normalized %>% 
  ungroup() %>%
  #filter(Metadata_id != "gn0055") %>%
  group_by(Metadata_matrix, Metadata_condition, Metadata_dose) %>%
  summarise(mean_speed = mean(Track_Speed),std_speed = sd(Track_Speed), n = n())
  
summary_speed %>% knitr::kable()
```

```{r}
ggplot(data = normalized %>% filter(Metadata_condition == 'CP17', Metadata_matrix == "FN"), aes(x = factor(Metadata_dose), y = Track_xFMI)) +
  geom_boxplot() +
  ggtitle("CP17 (normalized)")

ggplot(data = normalized %>% filter(Metadata_condition == 'control', Metadata_matrix == "FN"), aes(x = factor(Metadata_dose), y = Track_Speed)) +
  geom_boxplot( ) +
  ggtitle("Control (normalized)")

ggplot(data = normalized %>% filter(Metadata_condition == 'SP17', Metadata_matrix == "FN"), aes(x = factor(Metadata_dose), y = Track_Speed)) +
  geom_boxplot() +
  ggtitle("SP17 (normalized)")

ggplot(data = normalized %>% filter(Metadata_condition == 'vehicle', Metadata_matrix == "FN"), aes(x = factor(Metadata_dose), y = Track_xFMI)) +
  geom_boxplot( ) +
  ggtitle("Vehicle (normalized)")
```

We calculate the fration of tracks migrating in each direction. The migration in FN shows 44.2% neutros migrating towards the chemoattractant.
```{r}
sector_fraction <- normalized %>% 
  filter(Metadata_condition != 'water') %>%
  group_by_(.dots = strata) %>%
  summarise(positive = sum(Track_Positive_Sector) / n(),
            negative = sum(Track_Negative_Sector) / n(),
            neutral_up = sum(Track_Neutral_Sector_Up) / n(),
            neutral_down = sum(Track_Neutral_Sector_Down) / n()
    )

sector_fraction_condition <- sector_fraction %>%
  ungroup() %>%
  group_by(Metadata_matrix) %>%
  summarise(mean_positive = mean(positive),
            sd_positive = sd(positive),
            mean_negative = mean(negative),
            sd_negative = sd(negative),
            mean_neutral_up = mean(neutral_up),
            sd_neutral_up = sd(neutral_up),
            mean_neutral_down = mean(neutral_down),
            sd_neutral_down = sd(neutral_down)
            ) %>%
  print
```

Do we see a difference in the conditions? 

Summary: 
1. The migration in HEM does not show such a high variance as in FN (mean 44%, sd 16%) compard to HEM (mean 33%, sd 9% )
2. For HEM, the control 

```{r}
sector_fraction_condition <- sector_fraction %>%
  ungroup() %>%
  group_by(Metadata_matrix, Metadata_condition, Metadata_dose) %>%
  summarise(mean_positive = mean(positive),
            sd_positive = sd(positive),
            mean_negative = mean(negative),
            sd_negative = sd(negative),
            mean_neutral_up = mean(neutral_up),
            sd_neutral_up = sd(neutral_up),
            mean_neutral_down = mean(neutral_down),
            sd_neutral_down = sd(neutral_down)
            ) %>%
  print

x <- sector_fraction %>%
  filter(Metadata_condition == "vehicle", Metadata_matrix == "HEM")

y <- sector_fraction %>%
  filter(Metadata_condition == "CP17", Metadata_matrix == "HEM")

t.test(x$positive, y$positive)
```

```{r}
ggplot(data = normalized %>% filter(Metadata_matrix == 'HEM'), aes(x = Metadata_dose, y = Track_Speed)) +
  geom_boxplot( ) +
  ggtitle("Migration in HEM (normalized)")
```
First question: does the vehicle and control group show a significant difference in speed and directionality? 
```{r}
ttest_speed <- function(matrix, condition_x, condition_y){
  x <- normalized %>% 
    filter(Metadata_matrix == matrix) %>% 
    filter(Metadata_condition == condition_x)
  
  y <- normalized %>% filter(Metadata_matrix == matrix) %>% 
    filter(Metadata_condition == condition_y)

  return(t.test(x$Track_Speed,y$Track_Speed))
}

ttest_directionality <- function(matrix, condition_x, condition_y){
  x <- normalized %>% 
    filter(Metadata_matrix == matrix) %>% 
    filter(Metadata_condition == condition_x)
  
  y <- normalized %>% filter(Metadata_matrix == matrix) %>% 
    filter(Metadata_condition == condition_y)

  return(t.test(x$Track_Directionality,y$Track_Directionality))
}
```

```{r}
ttest_speed("HEM","SP17","CP17")
```

Using these function we can analyse the differences 
```{r}
t <-  ttest_speed("HEM","control","vehicle") %>% print
```
```{r}
t <-  ttest_speed("HEM","control","CP17") %>% print
```
# vehicle against control
```{r}
t <-  ttest_speed("HEM","vehicle","CP17") %>% print
```
# 
```{r}
t <-  ttest_speed("HEM","SP17","CP17") %>% print
```
CP17 and SP17 differ but not significantly. 
```{r}
x <- normalized %>% 
  filter(Metadata_matrix == "HEM") %>% 
  filter(Metadata_condition == "CP17") %>%
  filter(Track_Sector == 2) %>%
  filter(Metadata_dose == 50)
  
y <- normalized %>% 
  filter(Metadata_matrix == "HEM") %>% 
  filter(Metadata_condition == "SP17") %>%
  filter(Track_Sector == 2)   %>%
  filter(Metadata_dose == 50)

x %>% ungroup() %>% summarise(n = n())
y %>% ungroup() %>% summarise(n = n())
  

```

We see a difference in the Chemotactic Index
```{r}
t.test(x$Track_CI, y$Track_CI )
t.test(x$Track_xFMI, y$Track_xFMI )  
t.test(x$Track_Directionality, y$Track_Directionality )  
t.test(x$Track_Speed, y$Track_Speed )  
```
```{r}
normalized %>% 
  ungroup() %>% 
  filter(Metadata_condition == "vehicle") %>% 
  group_by_(.dots = strata) %>% slice(1) %>% 
  knitr::kable()
```
```{r}
hem <- normalized %>% 
  ungroup() %>% 
  filter(Metadata_matrix == 'HEM')

fn <- normalized %>% 
  ungroup() %>% 
  filter(Metadata_matrix == 'FN')

feature <-  c("Track_Speed","Track_Speed_Y","Track_Speed_X","Track_xFMI","Track_yFMI","Track_Directionality","Track_Distance_Traveled","Track_CI")

dif_fn <- DiffusionMap(fn, k = 50, vars = feature)

scatter3d(x = dif_fn$DC1, y = dif_fn$DC2, z = dif_fn$DC3, point.col = "blue", surface = FALSE, grid = FALSE, groups = factor( dif_fn$Metadata_condition ))
```



```{r}
ggplot(data = dif_fn, aes(x = DC1, y = DC3, color = factor(Metadata_condition))) +
  geom_point() + 
  labs(title = 'DM for migration in HEM')
  

ggplot(data = dif_fn, aes(x = DC1, y = DC3, color = factor(Metadata_id))) +
  geom_point() + 
  labs(title = 'DM for migration in HEM')

```

```{r}
normalized %>% 
  filter(Metadata_condition %in% c("vehicle")) %>%
  summarise( n = n()) %>% 
  filter(n < 40) %>% 
  arrange(Metadata_id) %>% 
  print
```
