==> 2017_06_track_matlab/05_find_subpopulation.Rmd <==
---
title: "03-data_visualization"
author: "TB"
date: "7/19/2017"
output: html_document
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library("fitdistrplus") 
library("readr")
library("tibble")
library("dplyr")
library("tidyr")
library("stringr")
library("ggplot2")

```

In this notebook we identify subpopulations in our data set based on the speed of the trajectories. We use the following approach: 

* we use tracking data normalized against the vehicle
* we estimate the distribution of the speed and calculate the .25 and .75 quantiles 
* these quantiles are used to define three groups of movement:
  * sneaking with speed < .25 quantile
  * walking with .25 quantile < speed < 

# Load data
We load the normalized and not normalized data.

```{r}
csv_tracks <- 'clean_tracks.csv'
csv_normalized <- 'clean_normalized.csv'
analysis_folder <- '../../../analysis/2017_04_28_matlab/'
strata <- c('Metadata_condition','Metadata_dose','Metadata_id','Metadata_date','Metadata_matrix')

normalized <- read_csv(file.path(analysis_folder,csv_normalized)) %>% 
  filter(Track_One_Cell == 1) %>%
  mutate(ID = str_c(Metadata_id, Metadata_condition, Metadata_dose, Metadata_matrix, sep = ' ')) %>%
  mutate(Metadata_condition = replace(Metadata_condition, Metadata_condition == "water", "vehicle")) %>%
  mutate(Track_Valid = Track_Length > 19) %>% 
  mutate(Track_Valid = as.numeric(Track_Valid)) %>% 
  print

population <- read_csv(file.path(analysis_folder,csv_tracks)) %>% 
  filter(Track_One_Cell == 1) %>%
  mutate(ID = str_c(Metadata_id, Metadata_condition, Metadata_dose, Metadata_matrix, sep = ' ')) %>%
  mutate(Metadata_condition = replace(Metadata_condition, Metadata_condition == "water", "vehicle")) %>%
  mutate(Track_Valid = Track_Length > 19) %>% 
  mutate(Track_Valid = as.numeric(Track_Valid)) %>% 
  print

```

Export mean, sd, sem and no. of valid tracks in each sector for all expriments
```{r}
population %<>% 
  mutate(Track_Sector = replace(Track_Sector, Track_Sector == 1, "positive sector")) %>%
  mutate(Track_Sector = replace(Track_Sector, Track_Sector == 2, "negative sector")) %>%
  mutate(Track_Sector = replace(Track_Sector, Track_Sector == 3, "neutral sector (up)")) %>%
  mutate(Track_Sector = replace(Track_Sector, Track_Sector == 4, "neutral sector (down)")) 

normalized %<>% 
  mutate(Track_Sector = replace(Track_Sector, Track_Sector == 1, "positive sector")) %>%
  mutate(Track_Sector = replace(Track_Sector, Track_Sector == 2, "negative sector")) %>%
  mutate(Track_Sector = replace(Track_Sector, Track_Sector == 3, "neutral sector (up)")) %>%
  mutate(Track_Sector = replace(Track_Sector, Track_Sector == 4, "neutral sector (down)")) 
```
  


Matrix HEM
```{r}
population %>% 
  ungroup() %>%
  filter(Track_Length > 19) %>%
  filter(Metadata_matrix == 'HEM') %>%
  ggplot(data = .) +
    aes(x = Track_Speed) +
    geom_histogram(bins = 100) + 
    scale_x_continuous(limits = c(0,20))

normalized %>%   
  filter(Track_Length > 19) %>%
  filter(Metadata_matrix == 'HEM') %>%
  ggplot(data = .) +
    aes(x = Track_Speed) +
    geom_histogram(bins = 100)

normalized %>%   
  filter(Track_Length > 19) %>%
  filter(Metadata_matrix == 'HEM', Metadata_condition == 'CP17') %>%
  ggplot(data = .) +
    aes(x = Track_Speed) +
    geom_histogram(bins = 75)

normalized %>%   
  filter(Track_Length > 19) %>%
  filter(Metadata_matrix == 'HEM', Metadata_condition == 'CP17', Metadata_dose == 50) %>%
  ggplot(data = .) +
    aes(x = Track_Speed) +
    geom_histogram(bins = 40)

normalized %>%   
  filter(Track_Length > 19) %>%
  filter(Metadata_matrix == 'HEM', Metadata_condition == 'CP17', Metadata_dose == 50) %>%
  summarise(n = n())

```
We visualize the normalized data to get an idea how the velocity is distributed. 
```{r}
population %>% 
  ungroup() %>%
  filter(Metadata_matrix == 'FN') %>%
  ggplot(data = .) +
    aes(x = Track_Speed) +
    geom_histogram(bins = 100) + 
    scale_x_continuous(limits = c(0,20))


normalized %>%   
  filter(Track_Length > 19) %>%
  filter(Metadata_matrix == 'FN') %>%
  ggplot(data = .) +
    aes(x = Track_Speed) +
    geom_histogram(bins = 100)


normalized %>%   
  filter(Track_Length > 19) %>%
  filter(Metadata_matrix == 'FN', Metadata_condition == 'CP17') %>%
  ggplot(data = .) +
    aes(x = Track_Speed) +
    geom_histogram(bins = 75)

  
normalized %>%   
  filter(Track_Length > 19) %>%
  na.omit() %>%
  filter(Metadata_matrix == 'FN', Metadata_condition == 'CP17', Metadata_dose == 50) %>%
  ggplot(data = .) +
    aes(x = Track_Speed) +
    geom_histogram(bins = 40)


normalized %>%   
  filter(Track_Length > 19) %>%
  filter(Metadata_matrix == 'FN', Metadata_condition == 'CP17', Metadata_dose == 50) %>%
  summarise(n = n())

```
```{r}
normalized %>% 
  mutate(Fast_Cell = as.numeric(Track_Speed > 0))
```


Proportion of valid tracks in each sector
```{r}
valid_tracks <- population %>% 
  group_by(ID,Track_Sector) %>%
  na.omit() %>%
  summarise(Fraction_valid_tracks = sum(Track_Valid) / n()) %>% 
  print

valid_tracks_all <- population %>% 
  group_by(ID) %>%
  na.omit() %>%
  summarise(Fraction_valid_tracks = sum(Track_Valid) / n()) %>% 
  mutate(Track_Sector = "all") %>%
  select(ID, Track_Sector, everything(.)) %>%
  print

valid_tracks <- bind_rows(valid_tracks,valid_tracks_all) %>% 
  ungroup() %>%
  group_by(ID) %>% print
```




# Define subpopulation based on the speed distribution. 

First, we filter the data. We only parameterize the migration run in Fibronectin.
```{r}
fn <- normalized %>%
  filter(Track_Length > 19, Metadata_matrix == "FN")
```


Is the data normal distributed?
```{r}
f1 <- fitdist(fn$Track_Speed,"norm") 
plot(f1)
shapiro.test(fn$Track_Speed) %>% print()
```
We want to quantify how many tracks are faster and smaller than the fastest and slowest 95% all all cells. 
```{r}
speed_quantiles <- quantile(fn$Track_Speed, probs = seq(0, 1, 0.25), na.rm = FALSE,
         names = TRUE, type = 7) %>% print

fn %<>% 
  mutate(sneaking = as.numeric(Track_Speed < speed_quantiles[2])) %>%
  mutate(walking  = as.numeric( (Track_Speed > speed_quantiles[2]) & (Track_Speed < speed_quantiles[4]) )) %>%
  mutate(running   = as.numeric(Track_Speed > speed_quantiles[4])) 



```

```{r}
subpopulation_speed <- fn %>% 
  group_by(Metadata_id, Metadata_condition, Metadata_dose,Track_Sector) %>%
  na.omit() %>%
  summarise(sneaking = sum(sneaking) / n(), walking = sum(walking) / n(), running = sum(running) / n()) %>% 
  print


subpopulation_speed_all <- fn %>% 
  group_by(Metadata_id, Metadata_condition, Metadata_dose) %>%
  na.omit() %>%
  summarise(sneaking = sum(sneaking) / n(), walking = sum(walking) / n(), running = sum(running) / n()) %>% 
  mutate(Track_Sector = "all") %>%
  dplyr::select(Metadata_id, Metadata_condition, Metadata_dose, Track_Sector, everything())


subpopulation_speed <- bind_rows(subpopulation_speed,subpopulation_speed_all) %>%
  arrange(Metadata_id, Metadata_condition, Metadata_dose, Track_Sector) %>%
  print

write_csv(subpopulation_speed, file.path(analysis_folder, 'subpopulation.csv'))
```
```{r}
subpopulation_speed %>% 
  ungroup() %>% 
  group_by(Metadata_condition, Track_Sector) %>%
  summarise(mean_sneaker = mean(sneaking), sd_sneaker = sd(sneaking), 
            mean_walking = mean(walking), sd_walking = sd(walking),
            mean_running = mean(running), sd_running = sd(running))
```
```{r}
p1 <- subpopulation_speed %>% 
  filter(Track_Sector == "positive sector") %>%
    ggplot(data = .) +
    aes(Metadata_condition, sneaking) + 
    geom_boxplot()

subpopulation_speed %>% 
  filter(Track_Sector == "positive sector") %>%
    ggplot(data = .) +
    aes(Metadata_condition, walking) + 
    geom_boxplot()

subpopulation_speed %>% 
  filter(Track_Sector == "positive sector") %>%
    ggplot(data = .) +
    aes(Metadata_condition, running) + 
    geom_boxplot()

subpopulation_speed %>% 
  filter(Track_Sector == "positive sector") %>%
    ggplot(data = .) +
    aes(Metadata_condition, sneaking) + 
    geom_boxplot()
```
```{r}
subpopulation_speed %>% 
  filter(Track_Sector == "positive sector") %>%
  filter(Metadata_dose == 50) %>%
  group_by(Metadata_condition) %>%
  print()

subpopulation_speed %>% 
  filter(Track_Sector == "positive sector") %>%
  filter(Metadata_dose == 50) %>%
  filter(Metadata_id != 'gn0017') %>%
    ggplot(data = .) +
    aes(Metadata_condition, sneaking) + 
    geom_boxplot()

s1 <- subpopulation_speed %>% 
  filter(Track_Sector == "positive sector") %>%
  filter(Metadata_id != 'gn0017') %>%
  filter(Metadata_dose == 5) %>%
  filter(Metadata_condition == "CP17") %>%
  print

s2 <- subpopulation_speed %>% 
  filter(Track_Sector == "positive sector") %>%
  filter(Metadata_id != 'gn0017') %>%
  filter(Metadata_dose == 5) %>%
  filter(Metadata_condition == "CP17") %>%
  print

kruskal.test(s1$running, s2$running)
```


```{r}
test <- fn %>% select(Metadata_id, Metadata_condition, Metadata_dose, running, walking, sneaking) %>%
  print

gather(test, pace, value, running)
```
